+++ src/persistence/common.jl
@@ -172,3 +172,10 @@ function getchild(el::EzXML.Node, s::String)
     i == 0 && error("no child $s")
     return childs[i]
 end
+
+"adds a child `s` to `el` if it doesn't exist"
+function getchild!(el::EzXML.Node, s::String)
+    childs = elements(el)
+    i = findfirst(x->name(x)==s, childs)
+    return i > 0 ? childs[i] : addelement!(el, s)
+end